@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "npm-run-all --parallel watch server app",
     "devweb": "npm-run-all --parallel server watch",
-    "server": "node nodemon_server.js",
+    "server": "node watchserver.mjs",
     "start": "neu run --frontend-lib-dev",
     "app": "neu run --frontend-lib-dev",
     "watch": "webpack --watch",
@@ -19,22 +19,21 @@
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.2",
-    "chalk": "^5.0.0",
+    "chalk": "^5.0.1",
     "classnames": "^2.3.1",
     "cors": "^2.8.5",
-    "dayjs": "^1.10.7",
+    "dayjs": "^1.10.8",
     "express": "^4.17.3",
     "hyperbee": "^1.7.0",
     "hypercore": "^9.12.0",
     "hyperdrive": "^10.21.0",
     "hyperspace": "^3.19.0",
-    "mine": "^0.1.0",
     "multer": "^1.4.4",
     "nanoid": "^3.3.1",
     "nodemon": "^2.0.15",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-router-dom": "^6.2.1"
+    "react-router-dom": "^6.2.2"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
@@ -43,10 +42,10 @@
     "@babel/preset-react": "^7.16.7",
     "@babel/runtime": "^7.17.2",
     "babel-loader": "^8.2.3",
-    "css-loader": "^6.6.0",
+    "css-loader": "^6.7.1",
     "npm-run-all": "^4.1.5",
     "style-loader": "^3.3.1",
-    "webpack": "^5.69.1",
+    "webpack": "^5.70.0",
     "webpack-cli": "^4.9.2"
   }
 }
