@@ -1,8 +1,15 @@
-const nodemon = require('nodemon');
-const path = require('path');
+/*
+  LICENSE: MIT
+  Created by: Lightnet
+*/
+
+import path from 'path';
+import nodemon from 'nodemon';
+import { URL } from 'url';
+const __dirname = new URL('.', import.meta.url).pathname;
 
 nodemon({
-  script:'./src/server/server.js'
+  script:'./src/server/server.mjs'
   //, stdout: false // important: this tells nodemon not to output to console
   , ext: 'js html'
   , env: { 'NODE_ENV': 'development' }
